UCS/TEST: Add strided allocator data structure
It allows allocating objects in a strided (instead of contiguous) memory
layout, which improves cache locality when some strides are used more
than others. The gap between the strides is compile-time constant for
faster access to the strides.
In addition, in order to distribute strided_alloc.h, need to move
UCS_PTR_BYTE_OFFSET to compiler_def.h